refact : LadderCreator 인터페이스 추가 및 Game이 DI 받도록 수정
