+++ 이현희/untitled/src/main/java/org/example/ver2/User.java
@@ -4,9 +4,6 @@ public class User {
     public void makeLadder(Game game, int width, int height) {
         game.makeLadder(width, height);
     }
-    public void makeLadderRandomly(Game game, int width, int height) {
-        game.makeLadderRandomly(width, height);
-    }
     public boolean drawHorizontalLine(Game game, int row, int column) {
         return game.drawHorizontalLine(row, column);
     }
